ajout pages modification, ajouts créature, ajout magie, restrcition d'accs, connexion deconnexion

+++ delete_creature.php
@@ -1,9 +1,25 @@
 <?php
-include_once 'environnement.php';
-$articleId = htmlentities($_GET['id']);
+    include_once 'environnement.php';
 
-$requestDEL = $bdd -> prepare('DELETE FROM monster
-                                WHERE id = ? ');  
-$requestDEL -> execute(array($articleId));
-header('Location: ./creatures.php?success=3');
-exit();
+    $request2 = $bdd->prepare('SELECT *
+                                FROM user
+                                INNER JOIN monster
+                                ON user.id = monster.user_id
+                                WHERE user_id = ?');
+    $request2 -> execute(array($sessionUserId));
+
+while ( $monsters = $request2->fetch()){
+
+    if($sessionUserId = $monster['user_id']){
+
+    $articleId = htmlentities($_GET['id']);
+
+    $requestDEL = $bdd -> prepare('DELETE FROM monster
+                                    WHERE id = ? ');  
+    $requestDEL -> execute(array($articleId));
+    header('Location: ./creatures.php?success=3');
+    exit();
+    } else {
+        header('Location: indexphp');
+    }
+}
